gemma2-27b-instruct/pyproject.toml
[project]
name = "gemma2-27b"
description = "Self-host google/gemma-2-27b-it with vLLM and BentoML"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "BentoML Team", email = "contact@bentoml.com" }]
dependencies = [
  "bentoml>=1.3.20",
  "vllm==0.7.1",
  "kantoku>=0.18.1",
  "openai>=1.61.0",
  "pyyaml",
  "Pillow",
  "flashinfer-python",
]
version = "0.0.0"

[project.urls]
Website = "https://bentoml.com"
Documentation = "https://docs.bentoml.com"
GitHub = "https://github.com/bentoml/BentoVLLM"
Twitter = "https://twitter.com/bentomlai"
Tracker = "https://github.com/bentoml/BentoVLLM"

[tool.bentoml.build]
service = "service:VLLM"
include = ["LICENCE", "*.py", "*.toml", "*.md"]
